Alterar propriedades de float para double em leituras
As propriedades `AccelX`, `AccelY`, `AccelZ` e `Temperature`
nas classes `SensorReadingDto` e `SensorReading` foram
modificadas de `float` para `double`. Essa mudança visa
aumentar a precisão dos valores de leitura de sensores,
abrangendo tanto os valores obrigatórios quanto o opcional.

+++ SensorApi/Models/SensorReading.cs
@@ -4,10 +4,10 @@
     {
         public Guid Id { get; set; } = Guid.NewGuid();
         public Guid DeviceId { get; set; }
-        public float AccelX { get; set; }
-        public float AccelY { get; set; }
-        public float AccelZ { get; set; }
-        public float? Temperature { get; set; }
+        public double AccelX { get; set; }
+        public double AccelY { get; set; }
+        public double AccelZ { get; set; }
+        public double? Temperature { get; set; }
         public DateTime Timestamp { get; set; } = DateTime.UtcNow;
         public Device? Device { get; set; }
     }
